\name{leerSIECyL-package}
\alias{leerSIECyL-package}
\alias{leerSIECyL}
\docType{package}
\title{
\packageTitle{leerSIECyL}
}
\description{
\packageDescription{leerSIECyL}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{leerSIECyL}
\packageIndices{leerSIECyL}
   The SIECyL is the System of Statistical Information (Sistema de Información Estadística) of the govern of Castilla y León (Spain).
   Access to the data is possible through the page http://estadistica.jcyl.es, but also using a query API.
This package uses this API to access any of the modules, at the moment only Datos Básicos, Coyuntura and Padrón.
The three main functions are:
leerDB to obtain data from the Datos Básicos module
leerCoyuntura to do it from the Coyuntura module.
leerPadron to do it from the Padrón module.
}
\author{
\packageAuthor{leerSIECyL}

Maintainer: \packageMaintainer{leerSIECyL}
}

